Update the cached selected text if we return early in on_text_selection_changed()
In 40e6ff10, we began returning early for text-selection-changed events
whose probably cause could not be determined. The goal was/is to stop
unwanted chattiness. It turns out that returning early is not enough if
we subsequently get a text-caret-moved event for the same object, because
we also present selection changes in response to that event. To stop
that from occurring, update the cached selected text before returning
early. That way, if we subsequently get an event causing us to consider
presenting the change in selection, we'll conclude it hasn't changed
since we last presented the text object.