Add "search" as a TextEventReason
If focus is in an entry we believe to be a search field, present text-
selection-changed events associated with the search unless we have a
reason not to. Note that very few apps and toolkits give us what we
need to make this determination globally. Bugs will be filed and
documentation written to explain Orca's expectations.