Event Manager: Try to catch more event flooding from Caja
Caja is flooding us horribly, and the logic that should catch it was not
because the offending object is focused (so we err on the side of caution).
In order to allow the logic designed for such messes to do its job, check
that an object does not manage its descendants before assuming its focused
state is good enough to make the event unconditionally processable.