Web: Ignore caert-moved events in non-editable objects triggered by Tab
We were already doing so for documents and links, but depending on the
content the events can be fired by a parent or descendant. If we fail to
ignore them, we won't speak anything, but we might update braille in
response removing the focused object from the display. Therefore create a
new TextEventReason to identify caret-moved events resulting from a focus
change and ignore them for web content.