Don't treat radio button label as group label if name is the same
In some applications and toolkits, the radio button group label is
exposed as a label for the radio button. In other toolkits, every radio
button has its own label which is the source of the name. The latter needs
to be identified and not confused for the radio button group label.

In addition, because some toolkits don't use either technique, looking for
the named panel or filler as the group can lead to false positives if we
ascend too high. The only known instance of this is in adwaita/GTK4 lists.
So stop if we reach a list before reaching a panel/filler.