CI: Undo wildcard change for tarball artifact path; start using x.y.z tags
The previous commit solved the problem of the tarball not being found.
However, the release-service job doesn't like wildcards. The path of
least resistance appears to be to stop using our tagging scheme and to
start tagging with the release version (i.e. matching what we use for
meson versions).