Check items for the selectable state instead of checking the parent's iface
When determining whether or not to say "not selected", we were checking if
the parent object supported the selection interface. If it did not, then we
assumed the children were not selectable. We should instead check if the
item itself has the selectable state. If the item is not selectable, that
state should be absent.

See https://gitlab.gnome.org/GNOME/gtk/-/issues/7118.