keybindings: Fix binding keybindings with the Shift key when using keysyms
Keysyms are representing characters, not keyboard positions,
so they are case sensitive. Account for that when binding shortcuts with the
shift key present in its modifier mask.

We are getting all key presses, including modifiers, regardless of what
keysym is the result of the given modifier mask and key code, so we are not
using a different modifier mask for the second key definition.