Web: Ignore caret-moved events from children-presentational roles
Example: When toggling the APG ARIA switch example, Gecko fires caret-moved
events for descendant text elements because... who knows? If we treat these
events as valid, we will update the current location and stop presenting
updates to the switch. To fix this, add a children_are_presentational()
utility based on the roles which meet this definition according to the ARIA
spec. Then, if a descendant of one of those roles fires a caret-moved event,
ignore it.