Work around Brave popup menus resulting in no window with active state
When the Brave popup menus appear, the browser frame fires window deactivate.
No window fires window activate. And no window has the active state. Because
we don't know what window the user is in, we clear the active window and wait
for a window activation or focused changed event so we can update the current
location and active script. Unfortunately those events fail to be fired until
the popup menu closes. As a result, we're not presenting the selection changes
in the popup menus.

To work around this issue, only clear the active window if we are able to
determine what window is now active. If no window claims to be active,
assume we're still in the current app for the purpose of presenting events.