GNOME Shell: Ignore focus claim from panel ancestor of current focus
From a user's debugging output, it appears that GNOME Shell is firing a
focus event for the newly focused widget (e.g. a password field) and then
firing a focus event for a panel parent/ancestor of that widget. If we
process the event for the panel, we'll cut off speech and update our
location incorrectly.