soffice: Don't announce "blank" as text for non-text objects
While "blank" may be useful as an information that a
text object (like a LibreOffice Writer paragraph)
does not contain any text, it's not useful for non-text
objects. Therefore, only try to get the text content
of the current line and return "blank" if it's empty
in the soffice SpeechGenerator._generate_text_line method
if the object supports the AtSpiText interface.

Otherwise, this causes the

    result += (self._generate_text_line(obj, **args) or \
        self._generate_accessible_label_and_name(obj, **args))

logic in SpeechGenerator._generate_list_item to use
"blank" for non-text items instead of their accessible label/name.

Sample scenario:

1) start LibreOffice Writer (gtk3 or qt6 variant,
   tested with current LO development version as of
   commit 8b998ed161f0c7e72530e5f02aa9be13edb4f703)
2) open the "Format" -> "Paragraph" dialog
3) go to the "Borders" tab
4) move focus to the "Shadow Style" section, list of
   items labelled by the "Position" label
5) move between the items using the arrow keys

Result without this commit in place:
Orca announces "blank" for each of the items.

Result with this commit in place:
Orca announces the actual name
("None", "Cast Shadow to Bottom Right",...).