Web: Ignore focus claim from ancestor dialog/alert
When navigating in the Discord inbox, we are sometimes getting a focus
event from the dialog containing our current location. If we update
our current location in response, we'll loop back to the top of the
list. Therefore, ignore focus claims from ancestor dialogs/alerts if
they are fired from an ancestor of the current location.

See issue #559.