Web: Handle object destruction triggered by scrolling element into view
In (at least) certain scrollable lists, scrolling the item into view
can cause the destruction and re-creation of the object we just scrolled
to. This in turn causes Orca to say "blank" and possibly loop to the
beginning of the list during structural navigation. Add a check to see
if the object may have been destroyed after the scroll and if so, try
to find its replacement based on name and role.

See issue #555.