LibreOffice: Remove frameAndDialog() override
frameAndDialog() was overridden for the purpose of speaking the sheet
name as part of the title in Where Am I for Calc. But returning the
spreadsheet as a dialog causes flat review to be restricted to the
sheet. A cleaner solution is to add the sheet-name support to the
generator.