Web: Speak/skip blank lines based on the "speak blank lines" setting
In order to provide efficient navigation, we were skipping blank lines
in web content when that content wasn't code. But a user reported wanting
to hear blank lines in web content. Since we have a "speak blank lines"
setting, use that as the basis to determine whether a blank line is
presented or skipped.