Focus locus of focus update when name of focused object changes
We were presenting name changes to the focused object via the presentation
of a message. This works for speech, and initially for braille. But because
a flash message is used, the braille displayed returns to the previous
contents (in this case the old/stale name) after a timeout. Fix that by
forcing an update to the locus of focus.