Remove the ancient J2SE script
This script consisted of ancient hacks for accessibility bugs in Java.
Some of these bugs may have since been fixed in the Java ATK Wrapper.
Identifying other bugs that can hopefully be similarly fixed will be easier
to do with this script out of the way. As we find issues, we can address them
either in JAW, or through less hacky solutions in Orca.