Prefer provided startOffset when generating text indentation
When generating text indentation, we were obtaining the indentation
for the current line as defined by the line with the caret. In web
content, if native navigation is disabled, the offset reported by the
user agent might be -1 or 0. That, in turn will cause us to report
the indentation of the first line rather than the actual line.

Because we include the start offset of the current line when
generating the text of web content, we should prefer that offset.