Remove internal content-error role
aria-invalid used to be supported on every ARIA role. That stopped being
the case a couple of years ago. We already present the invalid state for
widgets, but needed a way to check all non-widget roles just in case an
author set aria-invalid on a text-block element role. Doing this check
impacts performance. And now that it's no longer supported in ARIA, it
makes sense to stop doing this work.