Web: Fix presentation of value of readonly spin buttons
The logic which attempts to identify "helpful" offscreen text that
some authors provide -- and which is exposed to us unhelpfully as
single-character lines -- examines non-editable text objects only.
We need to also check if the object is a text input to prevent
falling back on the accessible name incorrectly.

Fixes issue #585.