Flat review: Don't create a flat review context when presenting a character
Creating the flat review context can be non-performant. If the user is
asking for a single character, and we're in a text object, and we are
not yet in flat review, get the character from the text object instead
of doing all the additional work to create the context. If we're already
in flat review do the normal behavior.

See issue #577.