Caret navigation: Fix presentation of next word for end of embedded child
If the "word" to the right consists of the content of the last word in
an embedded object followed by the space of the parent object, the
normal space-adjustment we do to set the caret at the end of the word
will cause us to erroneously present the first word of the embedded child
rather than the last. Fix that by removing the final content element
when it is a different object and consists only of whitespace.