Eliminate false positive in _is_layout_only()
In order to eliminate some chattiness with GTK4 apps, panels and groupings
inside set containers were treated as layout only. This was causing Orca
to not present ancestors it should, like labels in the preferences dialog.
Therefore only treat a named panel/grouping as layout only if it has the
same name as its set container.