Use generators to obtain name and text of cell headers
In the case of ARIA headers row/column headers, we're expected to
prefer the name. In Calc, however, the name of a given cell is
the coordinates therefore we need to prefer the text. We already
have handling for this situation when presenting the cell itself,
therefore apply that handling when generating cell headers.