Don't generate column header if the column header hasn't changed
The logic which generates changed row/column headers has filtering
to remove any headers that are shared between the old cell and the
new cell. That logic fails if obtaining a particular header cell
results in a new object being returned each time. See, for instance,
https://bugs.documentfoundation.org/show_bug.cgi?id=166418.

We weren't getting bitten by this failure for row headers because we
had a check to return early if the row hadn't changed. Such a check
makes sense for column headers, and also works around the LO bug.

Note, however, that returning early causes headers to not be presented
during whereAmI. Therefore check for that condition both for row and
for column headers