Cancel already-bound keybinding message in multi-click scenarios
Under certain circumstances, we are announcing the captured message
(e.g. "key captured Orca+T (double click)") and then stating that
the key was already bound due to the single-click binding. We need
to store the id from the call to GLib.timeout_add() and cancel the
message if it no longer applies.

See issue #596.