Structural Navigation: Handle weird authoring
If an author puts an ARIA heading inside a native heading, going to the
previous heading can fail because the caret context of the native heading
will be the ARIA heading. Thus going to the previous heading from the ARIA
heading will be the native heading. Handle this authoring by checking for
such nesting.

See issue #600.