Action Presenter: Use a GtkListBox instead of a GtkMenu
Using a GtkMenu for the Action Presenter means we have to deal with
GtkMenu possibly never popping up due to a timing issue related to
key grabs. See issue #410. There also appears to be some additional
weirdness specific to Wayland with this menu. Therefore stop using
a GtkMenu and use a GtkListBox in an undecorated window instead.