Web: Call shouldInterruptLForLocusOfFocusChange()
The default script's locus_of_focus_changed() checks to see if we should
interrupt speech before presenting the new location. This was done because
we were not getting input events for GTK4 apps and thus would never shut
up when the user's location changed. There are other scenarios in which
a web application might change focus without our having received an input
event. Therefore, add a check to the web script's locus_of_focus_changed()
so we can present that new location immediately.