Structural Navigation: Interrupt speech in goObject() if we lack an input event
If goObject() is called due to an input event such as "H" to go to the
next heading, speech will be interrupted due to the press of "H". If
navigation was not triggered by an input event, we still want to stop
speaking the previous location before presenting the new location.
Note that we cannot count on locus_of_focus_changed() in this case
because the structural navigation code "silently" updates the locus of
focus in order to prevent double-presentation (i.e. should the navigation
trigger the object to claim focus).