ci/build_and_install.sh: Use the system python, not the Python environment that is just for CI
The system's installation of Python has Orca's run-time
dependencies (brlapi, dasbus, louis, etc.) - those come from the
distro's packages.

However, type checking and linting of the Python code just happens in
the CI, and for that we use an environment in /usr/local/python (see
ci/install-python-tools.sh).  The reason to not use system packages is
that not all of them are available pre-packaged.

Part-of: <https://gitlab.gnome.org/GNOME/orca/-/merge_requests/248>