Fix early return in _generate_state_invalid()
_generate_state_invalid() checks for the text attributes at offset
and returns early if there is no error attribute or the error
attribute is set to false. That prevents the error message from being
presented for widgets which have an invalid-entry state but no text
(e.g. a checkbox) or no "error" text attribute (e.g. native inputs).

See issue #514.