Web: Enter focus mode when focus moves inside descendant of web app
If the author has set aria-activedescendant on an element with the
application role and a tabindex, user agents fire the focus event
for the descendant (which might not even be focusable). As a result,
the decision regarding whether or not to enter focus mode was being
based on whether or not the descendant should be a focus-mode widget.
Fix that by entering focus mode when focus moves into an element with
the application role.

Fixes issue #610.