Ignore collapsed events from objects we're not on/in
If an object that we were in fires a collapsed event and claims to be
focused (even if it's not), we should not present this event. This is
especially the case in web content where we foolishly trust that something
claiming to be focused is actually focused and update our location
accordingly. Therefore, if a collapsed event is fired and it's not the
locus of focus or an ancestor of the locus of focus (which could happen
in the case of a combo box), ignore it.