Remove unmaintained and broken mouseover code
This code was designed to trigger and present the appearance of
an element which only appears in response to mouse hovering. In
order for this to work, all of the following must be true:

1. The user somehow knows that the currently focused object has a
   hover action that will cause something to pop up.
2. The user was able to route the mouse pointer to the object via
   Orca command. This command doesn't work in Wayland. And in X it
   depends heavily on the object being scrolled into view and our
   having accurate bounding boxes for the object (both of which
   are often not the case).
3. As a result of successfully routing the mouse pointer to the
   object, not only did something pop up, but the user agent fired
   a children-added event for the container of the popup and did so
   before firing any other children-added events.
4. The popup remains visible on screen so its accessible object is
   not destroyed.

If there really is a need for this feature, and all of the above
issues have been addressed, we can re-add/implement this functionality
in the AXEventSynthesizer.