Structural Navigator: Present first child rather than list for L/Shift+L
If the list is large enough, it is more performant to present the list
as the ancestor of the first child. In addition, there is missing handling
when calculating the objects on the line for a list. This wasn't known
before the structural navigation refactor because we were moving to the
first child for list navigation. To fix the performance regression, restore
that behavior. Adding the missing handling will be done in a follow-up
change.

See issue #616.