Web: Defer to default script's presentObject() for list items
Prior to the structural-navigation refactor, we had very special handling
for lists, namely to set focus to the first item. As part of the refactor,
everything switched over to using presentObject() unconditionally. That
was super-non-performant for lists, so half of the very special handling
was restored in 41ab3bf244. But if we call presentObject() on a list item
with a huge nested list, we're still going to do a full dive into that
object. Therefore, prevent the web script's presentObject() from doing a
dive, similar to what we do for giant entries (e.g. VS Code).