Don't shutdown on screen-reader-enabled -> false
With systemd startup, this will no longer work: the only way to shut
down Orca correctly would be to stop and disable the systemd service.
This prevents weird edge-cases where the screen reader is supposed to be
running but isn't.

In GNOME, gnome-settings-daemon is responsible for managing the systemd
service in tandem with the gsetting.

MATE, Cinnamon, and Unity all use old versions of gnome-session or forks
of gnome-session. Either way, gnome-session's builtin service manager
will kill orca-autostart.desktop whenever the gsetting is false, so Orca
doesn't need to do this on its own.

KDE is currently using the gsetting to stop Orca, but I spoke to the KDE
developers and they were interested to use the systemd service directly,
like GNOME will be doing. This presents an opportunity to do some
cleanup for KDE, which is currently forking off `gsettings set` commands
to kill Orca