Handle speechServerInfo disconnect from speechServerFactory
settings.speechServerInfo, if set, is the [name, id] of the user's chosen
synthesizer. It has no connection the settings.speechServerFactory, which
is either speechdispatherfactory or spiel. As a result, if we switch from
one factory to another, the new factory is almost certainly looking at
invalid server info. Handle that scenario by checking if the server is
known to the factory and, if not, using the default server.