Perform a "health check" when getting the speech server
SpeechAndVerbosityManager's _get_server() was returning the server to
other callers assuming the server was alive and well. If the server
is hung, that means we'll potentially perform multiple blocking calls.
Therefore, add a "health check" (currently asking for the output
module). If that call times out, _get_server() will return None.