InputEventManager: Pause processing keyboard events when EventManager pauses
We don't process an incoming AtspiEvent when we're reloading settings or
shutting down. We should treat any incoming KeyBoardEvent in the same
fashion because we're not in a position to handle it (active script might
be changing, which implies refreshing keygrabs; speech server might be
changing or shutting down, so we shouldn't try to send it commands; etc.)

See issue #619.