systemd: Integrate with watchdog
This is a feature in systemd that ensures a service doesn't get stuck in
a non-responsive state.

Once the service tells systemd that it's running, systemd will start
expecting the service to send pings (via sd_notify) at regular
intervals. If the service misses a ping, systemd assumes that the
service got stuck stuck and aborts it. Since we have Restart=always
configured, after crashing Orca systemd will immediately restart a fresh
instance

This is an easy way to increase the robustness of Orca, which is a
critical service for people that need it