Work around hypertext/hyperlink implementation brokenness
We're seeing off-by-one errors in (at least) Chromium where the text
at a given offset is an embedded object character, but retrieving the
child at that offset returns None.

Create a utility to handle this condition by checking for the child
at the previous and next offsets. If there's a child, and its offset
in parent matches the desired offset, return that child. This assumes
that we can get a valid offset in the parent, but from some testing
with broken test cases it appears we can.