Revert "Don't shutdown on screen-reader-enabled -> false"
So apparently some people rely on the fact that they can run an
unmanaged copy of Orca, then hit alt+super+s (potentially twice) to tell
it to quit. This works, but only on accident: if gnome-session was doing
its job it would be super broken, as I'll explain in the rest of
the commit message. But this is behavior that was in place for an
extended amount of time, and people rely on it. I've adjusted the commit
to only reintroduce this behavior whenever Orca is running outside of
systemd's management --- under systemd Orca behaves correctly.

The core of the problem is that gnome-session (and thus derivatives like
mate-session and cinnamon-session) is just plain broken. Orca requests
to be auto-restarted in the .desktop file (via
X-GNOME-AutoRestart=true), but gnome-session simply fails to do that. It
has no code to do that. This bug manifests in two different ways when
you try to run an unmanaged build of Orca.

Situation 1: You run `orca --replace` while the screen reader is
running. This forcibly kills all other instances of Orca, including the
one managed by gnome-session. Essentially, from the perspective of
gnome-session, Orca crashed. Of course, gnome-session makes no attempt
to restart Orca and so your `orca --replace` command worked. But if
gnome-session was behaving correctly, it would try to restart `orca`.
This would fail (since another instance is already running), which would
be an unclean exit. Unclean exit = a second restart. This would get
gnome-session stuck in a loop trying to repeatedly restart Orca. systemd
has "restart rate limits" implemented to avoid situations like this,
but gnome-session is nowhere near that sophisticated

Situation 2: You run `orca` while the screen reader is off. Then you
press alt+super+s to try and shut down Orca. What this actually does is
turn _on_ the screen reader. gnome-session immediately tries to spawn
Orca, which fails because another instance is running already. Again: in
reality gnome-session gives up at this point. But if gnome-session was
behaving as documented and intended, it would get stuck in a loop
repeatedly trying to restart Orca.

I would advise against relying on this behavior, because it would only
take some bugfixes in mate-session or cinnamon-session to suddenly start
triggering those "stuck in an infinite retry loop" scenarios.

This reverts commit e3a704cfd855ff338405916c67577d8309c3abfd.

Fixes https://gitlab.gnome.org/GNOME/orca/-/issues/621