orca-autostart.desktop: Add back OnlyShowIn=GNOME
We removed this since the .desktop file shouldn't be used in GNOME 49
and newer. But it turns out that, some people run new versions of Orca
with stable versions of GNOME. Removing GNOME from the .desktop file
broke autostart in this situation

Luckily, Orca has since gained integration with systemd, and now sets
X-systemd-skip=true in the desktop file. This communicates to
gnome-session that this .desktop file has direct systemd integration,
and so gnome-session will ignore this file.

So we can now safely include GNOME in OnlyShowIn. This ensures that for
GNOME 48 and older, Orca 49 will continue working as before. And for
GNOME 49 and newer, Orca 49 will be launched via systemd directly, and
orca-autostart.desktop will be completely ignored.

Reverts 5dca8c2b6d9d13785b4e209cb3d76b7c52263096