InputEventManager: Stop using _all_names_for_key_code()
In a handful of cases we were looking up keycodes instead of using
keysyms to identify if a given key was pressed. That approach fails if
we don't get the expected keycode. See issue #606.

Most of the code is already using _last_key_and_modifiers(). Update the
remaining functions to do the same and delete _all_names_for_key_code().
Having done so, also delete _last_keycode_and_modifiers() which is now
unused.