Present static text for GtkList
Orca presents the displayed static text of ancestor containers, prefering
the description as described in the app developer docs. It was not doing
so for the list role because that role had traditionally been reserved
for non-gui/document-content.

In order to respect what is stated in the docs without breaking other
lists, treat GtkList as a gui list and present static text for gui lists.

See issue #625.