Present system messages at "some" rather than "no" punctuation
When speech-dispatcher's pre-processing is enabled, symbols that should
be spoken in an Orca system message are not. The reason is that we were
using the "none" level so that announcements would be spoken in a fluid
fashion. Setting system messages to the level of "some" causes relevant
symbols (e.g. percent) to still be spoken while still stripping out
more traditional punctuation symbols.

Fixes issue #623.