Delay key echo of printable keys in terminal until we have a text insertion
We have no reliable way to identify that the user is typing a password
in a terminal. We historically did two things to work around this:

1. Echo key releases instead of key presses when in a terminal (i.e.
   do the complete opposite of what we do everywhere else).
2. Check the text at and before the caret offset. If that character
   matched the key, we echoed the release.

This workaround had associated problems, namely making the echo
logic complicated and the presentation laggy in the terminal. For
that reason, as part of the recent typing echo refactor, the logic
was simplified a bit, namely always echo key presses. Local testing
showed (and continues to show) that the accessible text is updated
by the time we do the surounding-text check. Unfortunately end-user
testing has proven this is not always the case.

Rather than go back to the old, laggy logic, this commit eliminates
the surrounding-text check. Thus we now do the following when a
key press is received from the terminal:

* If the press is for a non-printable key, use the default logic.
* If the press is for a printable key, save the event and return
  early.
* When a text insertion event is received from the terminal, the
  terminal script calls echo_delayed_terminal_press(). If the
  accessible objects match and the inserted text matches, echo
  the saved event.

Local testing shows this works too. Let's see what the users find.

See issue #630.