When generating descendants, ignore anything in the labelled-by list
Given the wide variety of GtkList items we see in the wild, Orca
has logic which attempts to locate all the things which should be
presented (e.g. a descendant widget whose state should be announced).
This logic also attempts to filter out redundant descendants, e.g.
based on role, based on their content appearing in the item name,
etc. But apparently this filtering is not sufficient.

In order to maximize the likelihood that Orca continues to announce
what it should, while giving app developers a means to stop Orca
from presenting something it shouldn't, Orca will now ignore any
descendant in the item's labelled-by relation targets.

To see what is in the labelled-by relation target list for your app,
you can use the `tools/dump-tree.py` utility in Orca's repo.

Example: For this node

--> ATSPI_ROLE_LIST_ITEM (0x7fa73b1ff6c0) NAME: 'foo' DESCRIPTION: 'bar'
    PARENT: ATSPI_ROLE_LIST (0x7fa73b1ff400) INDEX IN PARENT: 7
    ATTRIBUTES: toolkit=GTK, keyshortcuts=
    RECT: X:346, Y:591, WIDTH:674, HEIGHT:62
    STATES: focusable sensitive showing visible
    RELATIONS: labelled-by: ATSPI_ROLE_LINK (0x7fa73b200bc0)

Orca will now ignore the link listed in "RELATIONS" when generating
descendants.