Fix "1 of 1" announcement for GTK4 collapsed combo boxes
When a combo box doesn't expose posinset and setsize, we attempt to
determine it from the items/members of that combo box. If we cannot
find the selected item, we can't know the position. If we cannot find
any items, we cannot know the set size. In this case, we should return
early and make no announcement.

See issue #631.