Call _get_functional_role() for all objects when generating speech/braille
The main reason for the existence of _get_functional_role() is to improve
presentation of ARIA roles in cases where we lacked a corresponding, or
more precise role. But it also gives us a way to easily support new, real
roles (e.g. Atspi's 'switch' role) without having to insert try/except
everywhere. Therefore move the use of _get_functional_role() out of the
web script and into the default speech and braille generators.

See issue #632.