Fix false positive in _should_verbalize_punctuation()
We treat the content of a <pre> element in the same way we do <code>,
namely verbalize punctuation unless the user has set the punctuation
level to NONE (doesn't want it) or ALL (synthesizer will do it for us).
This was done because we were seeing code snippets wrapped in <pre>, and
users reported that Orca should verbalize that punctuation automatically.

Unfortunately, plain text documents being viewed in Firefox have their
content wrapped in a <pre>. Add a check for this condition and do not
verbalize punctuation for plain text documents.