Web: Try to make deeply nested layout tables less time-consuming
Seen in the wild (a forwarded email): 15 levels of table nesting
just to display some text and a banner. Trying to process the contents
is not only sluggish but can cause us to miss our systemd ping deadline.
Try to make this less awful by:

* Treating tables and table rows as non-text so that we access their
  children directly rather than descending down to them via the
  hypertext and hyperlink interfaces.
* Returning early when building up a line and the next/previous object
  is, or descends from a different table row.
* Treating all table-related objects as being valid for holding the
  caret context so we do less work processing these items.

These changes do not appear to impact presentation of normal tables.