InputEventManager: Don't add grab id of 0 to grabbed bindings
A grab id of 0 indicates failure. When we have double/triple-click
bindings, the single-click binding will be registered first, and
subsequent attempts to register that same key combination will fail.
If we only have a double/triple-click, the grab works as expected.
Adding these duplicate bindings to the grabbed bindings we keep track
results in irrelevant debug spew.