Add means to notify the systemd manager that Orca is still alive
Event processing is currently is done on the GLib main loop; not
a worker thread. As a result, when we receive hundreds of accessible-
object events a second, we can miss the systemd watchdog ping deadline.

* Add notify_alive() to the systemd manager. It will send the ping
  if one is due.
* Have the event manager call notify_alive() at the beginning of
  _enqueue_object_event().

Because _enqueue_object_event() is the callback registered with AT-SPI2
for accessible-object events, if it is called we know that both AT-SPI2
and Orca are alive and should not be killed and restarted.

See issue #635.