Use Atspi.CollectionMatchType.ALL by default
As is the case for the other match/filter criteria already,
also use Atspi.CollectionMatchType.ALL by default
for attribute and role matching in
AXCollection.create_match_rule.

Using a match type of Atspi.CollectionMatchType.ANY
with an empty list of attributes would for example
break structural navigation to jump between headers
in LibreOffice Writer with the Qt 6 variant of
LibreOffice.

This is because the Qt AT-SPI Collection implementation
introduced in qtbase commit [1] considers
Atspi.CollectionMatchType.ANY as meaning "at least one"
attribute matches, which is always false in case
there is no attribute to match against.

(The GTK 3/ATK implementation seems to handle
this differently, but I don't think Orca should
rely on that behavior.)

Adjust callers that actually want to use
Atspi.CollectionMatchType.ANY to pass that
explicitly instead of relying on the previous
default.

Together with pending Qt change [2], this makes jumping
between headings in the Qt 6 variant of LibreOffice work.
(Currently limited to the part of the document visible
on screen, see LibreOffice bug [3].)

[1] https://code.qt.io/cgit/qt/qtbase.git/commit/?id=7a478ff96f27fbfc8ddfbc67ba62276580aa02b4
[2] https://codereview.qt-project.org/c/qt/qtbase/+/685892
[3] https://bugs.documentfoundation.org/show_bug.cgi?id=35652