Work around GTK3 radio button child index variability
For GTK3 radio buttons, we use the member-of relation to identify all
group members. Unfortunately the targets are not necessarily in order,
thus we sort them by child index. Often GTK3 radio buttons appear in
the accessibility tree in the reverse order (i.e. first radio button is
the last child). But that's not always the case (e.g. the radio buttons
in the Budgie panel are in the correct order).

To handle both scenarios, sort first by child index. Then compare the
spatial location of the first two. If the first is spatially after the
second, reverse the sorted list.

See gtk#7839.