Minimize unintentional voice switches and pauses
* Only attempt to determine language + dialect if we don't have the
  language.
* Only switch language + dialect if the language has changed.

If just the dialect has changed, we can wind up with valid dialect
switches (e.g. between American and British English) which may be
undesirable for some users. This can become an opt-in feature later.

We can also wind up with pauses between utterances when either the
user's setting or the provided value is missing a dialect. This is
the result of the family definitions not being an exact match which
causes each utterance to be treated as a separate spoken item.