Gather and use fallback voices when there's no exact match
If the content provides a language but no dialect, there will not be
an exact match if all voices of that target language do expose a dialect.
Therefore, gather plausible fallback voices. Plausible is currently defined
having a two-letter dialect code (e.g. GB, US vs. GB-X-RP, en-US-NYC) so
that *if* we cannot find an exact match, we're more likely to use a more
commonly heard dialect of the desired language.