Say All Presenter: Prevent focus-mode/browse-mode switches
In web content, Say All can be "complete" according to the speech server
because all of the current content has been presented. But in web content
we progressively work our way through the document and might not really be
complete. In order to prevent unwanted focus-mode/browse-mode switches,
keep track of whether or not we think say all is running by setting it to
True in the progress callback and setting it to False if Say All is
interrupted as well as at the end of _say_all_iter.

See issue #645.