Use settings.fooBar instead of SettingsManager.get_setting("fooBar")
The ultimate goal is for each manager/presenter/navigator to manage its
own settings, rather than having a global manager, global file, etc.
Given the size of that change, the work is being done in smaller chunks.
This commit replaces use of SettingsManager's get_setting with direct
access to the setting via settings.py. Given that get_setting() is simply
a wrapper for that direct access, this change should not impact behavior.